requirements.txt
blinker==1.4
boto3==1.9.199
botocore==1.12.199
certifi==2019.6.16
Click==7.0
docutils==0.14
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
mysql-connector==2.2.9
python-dateutil==2.8.0
s3transfer==0.2.1
sentry-sdk==0.11.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
WTForms==2.2.1
